requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.34,<1.0
openpyxl>=3.1,<4.0
xlrd>=2.0,<3.0
python-multipart>=0.0.18,<1.0
pydantic>=2.10,<3.0
PyJWT>=2.10,<3.0
passlib[bcrypt]>=1.7,<2.0
python-dotenv>=1.0,<2.0
google-cloud-storage>=2.18,<3.0
